Add files via upload
The Revision 2 uses a combination of 2022 and 2024 data to train and test decision trees. The testing data is not filtered, so it should contain all attack types and benign data to allow the decision tree to identify if an attack is present and whether it is of the specific attack type for that model. Further revisions will prove this to a higher extent and expand on testing with various training ranges, data splits, and orders.