Add files via upload
This revision trains on dataset 6 which contains only benign data. The forecast generated by the ARIMA model is compared to a test dataset which contains a combination of dataset 7 and dataset 8 which contains only attack data. The test data and forecasted data are sent to a final dataframe which can be used for further testing.